fix: Resolve circular dependency between AI packages preventing builds (#1053)
* fix(AI): Resolve circular dependency between aiengine and ai-prompts packages

Fixes circular dependency that was preventing MJCoreEntitiesServer from building:
- AI/Engine imported AIPromptRunResult from AI/Prompts
- AI/Prompts depended on AI/Engine in package.json

Changes:
- Move AIPromptRunResult and related types to @memberjunction/ai package
- Update AI/Engine to import AIPromptRunResult from @memberjunction/ai
- Update AI/Prompts to import types from @memberjunction/ai instead of defining locally
- Add comprehensive build failure debugging guide to CLAUDE.md

This resolves the build failure where MJCoreEntitiesServer couldn't build during npm install
due to the circular dependency blocking the aiengine prerequisite.



* fix(AI/Prompts): Update imports after moving types to @memberjunction/ai

Updates AIPromptRunner.ts to import moved types from @memberjunction/ai instead of local types file.
This completes the circular dependency fix by ensuring all type references point to the shared location.

Types moved: ExecutionStatus, CancellationReason, ModelInfo, JudgeMetadata, ValidationAttempt, AIPromptRunResult



* fix(AI/Agents): Update imports after moving types to @memberjunction/ai

Updates AI Agents package to import AIPromptRunResult from @memberjunction/ai instead of
@memberjunction/ai-prompts, completing the circular dependency resolution.

Files updated:
- base-agent-type.ts: Import AIPromptRunResult from @memberjunction/ai
- loop-agent-type.ts: Import AIPromptRunResult from @memberjunction/ai
- base-agent.ts: Import AIPromptRunResult from @memberjunction/ai



* docs(changeset):   **Breaking circular dependency between AI packages**

* changeset description

---------
