+++ packages/AI/Agents/src/agent-types/loop-agent-type.ts
@@ -13,7 +13,7 @@
 import { RegisterClass } from '@memberjunction/global';
 import { BaseAgentType } from './base-agent-type';
 import { BaseAgentNextStep } from '@memberjunction/aiengine';
-import { AIPromptRunResult } from '@memberjunction/ai-prompts';
+import { AIPromptRunResult } from '@memberjunction/ai';
 import { LogError, LogStatus } from '@memberjunction/core';
 
 /**
